backend_django/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.2.2
xgboost==1.7.5
joblib==1.2.0
python-dotenv==1.0.0
